api/requirements.txt
Flask==2.1.0
Pillow==9.3.0
pymongo==4.1.1
sentence_transformers==2.2.2
Werkzeug==2.2.2